requirements.txt
python-telegram-bot[job-queue]>=22.0
openai>=1.0.0
anthropic>=0.45.0
python-dotenv>=1.0.0
pydantic>=2.0.0
